+++ packages/server-common/src/handlers/calendar/index.ts
@@ -1,3 +1,3 @@
 export * from './calendar-handler-factory.contentful';
-export * from './resubscribe-handler';
+export * from './unsubscribe-handler';
 export * from './subscribe-handler';
