+++ packages/server-common/src/handlers/webhooks/events-updated-process.handler.ts
@@ -40,6 +40,11 @@ export const webhookEventUpdatedProcessHandlerFactory = (
       const resourceId = record.messageAttributes.ResourceId?.stringValue;
       const channelId = record.messageAttributes.ChannelId?.stringValue;
 
+      if (!resourceId) {
+        logger.debug('Empty resourceId');
+        return;
+      }
+
       if (!(resourceId && channelId)) {
         throw new Error('Invalid payload');
       }
